Rename FormComponent to Form
