--- src/components/FormComponent/index.ts
@@ -1,1 +1,0 @@
-export * from './FormComponent';
